split app into files
