+++ src/state.js
@@ -1,0 +1,9 @@
+/* @flow */
+import type { State } from './types';
+
+export default function createState(): State {
+  return {
+    selected: null,
+    polygons: [],
+  };
+}
